Fix gobject hook after utils.hooks refactor.
