+++ PyInstaller/hooks/hook-gi.repository.Atk.py
@@ -11,7 +11,8 @@
 """
 Import hook for PyGObject https://wiki.gnome.org/PyGObject
 """
-from PyInstaller.utils.hooks import collect_glib_translations, get_gi_typelibs
+from PyInstaller.utils.hooks.gi import collect_glib_translations, \
+        get_gi_typelibs
 
 binaries, datas, hiddenimports = get_gi_typelibs('Atk', '1.0')
 datas += collect_glib_translations('atk10')
