+++ PyInstaller/hooks/hook-gi.repository.Clutter.py
@@ -12,6 +12,6 @@
 Import hook for PyGObject's "gi.repository.Clutter" package.
 """
 
-from PyInstaller.utils.hooks import get_gi_typelibs
+from PyInstaller.utils.hooks.gi import get_gi_typelibs
 
 binaries, datas, hiddenimports = get_gi_typelibs('Clutter', '1.0')
