+++ PyInstaller/hooks/hook-gi.repository.Gio.py
@@ -23,7 +23,7 @@ import sys
 
 import PyInstaller.log as logging
 from PyInstaller.compat import is_darwin, is_win, is_linux, base_prefix
-from PyInstaller.utils.hooks import get_gi_typelibs, get_gi_libdir, exec_statement
+from PyInstaller.utils.hooks.gi import get_gi_typelibs, get_gi_libdir
 
 logger = logging.getLogger(__name__)
