+++ PyInstaller/hooks/hook-gi.repository.Gtk.py
@@ -17,7 +17,8 @@ import os.path
 import glob
 
 from PyInstaller.compat import is_win
-from PyInstaller.utils.hooks import collect_glib_share_files, collect_glib_etc_files, collect_glib_translations, exec_statement, get_gi_typelibs
+from PyInstaller.utils.hooks.gi import collect_glib_share_files, \
+        collect_glib_etc_files, collect_glib_translations, get_gi_typelibs
 
 binaries, datas, hiddenimports = get_gi_typelibs('Gtk', '3.0')
