+++ PyInstaller/hooks/hook-gi.repository.GtkChamplain.py
@@ -12,6 +12,6 @@
 Import hook for PyGObject's "gi.repository.GtkChamplain" package.
 """
 
-from PyInstaller.utils.hooks import get_gi_typelibs
+from PyInstaller.utils.hooks.gi import get_gi_typelibs
 
 binaries, datas, hiddenimports = get_gi_typelibs('GtkChamplain', '0.12')
