+++ PyInstaller/hooks/hook-gi.repository.GtkClutter.py
@@ -12,6 +12,6 @@
 Import hook for PyGObject's "gi.repository.GtkClutter" package.
 """
 
-from PyInstaller.utils.hooks import get_gi_typelibs
+from PyInstaller.utils.hooks.gi import get_gi_typelibs
 
 binaries, datas, hiddenimports = get_gi_typelibs('GtkClutter', '1.0')
