+++ PyInstaller/hooks/hook-gi.repository.GtkSource.py
@@ -10,7 +10,8 @@
 #-----------------------------------------------------------------------------
 
 
-from PyInstaller.utils.hooks import collect_glib_share_files, get_gi_typelibs
+from PyInstaller.utils.hooks.gi import collect_glib_share_files, \
+        get_gi_typelibs
 
 binaries, datas, hiddenimports = get_gi_typelibs('GtkSource', '3.0')
