+++ PyInstaller/hooks/hook-gi.repository.GtkosxApplication.py
@@ -13,7 +13,7 @@ Import hook for PyGObject https://wiki.gnome.org/PyGObject
 """
 
 from PyInstaller.compat import is_darwin
-from PyInstaller.utils.hooks import get_gi_typelibs
+from PyInstaller.utils.hooks.gi import get_gi_typelibs
 
 
 if is_darwin:
