+++ PyInstaller/hooks/hook-gi.repository.HarfBuzz.py
@@ -12,7 +12,7 @@
 Import hook for PyGObject https://wiki.gnome.org/PyGObject
 """
 
-from PyInstaller.utils.hooks import get_gi_typelibs
+from PyInstaller.utils.hooks.gi import get_gi_typelibs
 
 
 binaries, datas, hiddenimports = get_gi_typelibs('HarfBuzz', '0.0')
