+++ PyInstaller/hooks/hook-gi.repository.Pango.py
@@ -12,6 +12,6 @@
 Import hook for PyGObject https://wiki.gnome.org/PyGObject
 """
 
-from PyInstaller.utils.hooks import get_gi_typelibs
+from PyInstaller.utils.hooks.gi import get_gi_typelibs
 
 binaries, datas, hiddenimports = get_gi_typelibs('Pango', '1.0')
